+++ projects/packages/videopress/src/client/admin/components/video-quick-actions/types.ts
@@ -17,6 +17,14 @@ export interface ActionItemProps extends React.ButtonHTMLAttributes< HTMLButtonE
 
 export interface PopoverWithAnchorProps {
 	/**
+	 * Whether the popover should be rendered or not
+	 */
+	showPopover?: boolean;
+	/**
+	 * Whether the anchor is focused
+	 */
+	isAnchorFocused?: boolean;
+	/**
 	 * Ref that anchors the popover
 	 */
 	anchor: HTMLElement | null;
@@ -69,6 +77,7 @@ export type ThumbnailActionsDropdownProps = {
 	onUpdate: ( action: 'default' | 'select-from-video' | 'upload-image' ) => void;
 	description: string;
 	isUpdatingPoster?: boolean;
+	disabled?: boolean;
 };
 
 export type PrivacyActionsDropdownProps = {
@@ -76,4 +85,5 @@ export type PrivacyActionsDropdownProps = {
 	privacySetting?: privacySetting;
 	isUpdatingPrivacy?: boolean;
 	description: string;
+	disabled?: boolean;
 };