+++ projects/packages/videopress/src/client/block-editor/blocks/video/components/videopress-player/index.js
@@ -94,9 +94,13 @@ export default function VideoPressPlayer( {
 			return;
 		}
 
+		// Once the video is loaded, delegate the height to the player (iFrame)
 		if ( preview ) {
-			// Once the video is loaded, delegate the height to the player (iFrame)
-			return setVideoPlayerTemporaryHeightState( 'auto' );
+			// Hack to mitigate the flickr when the player is
+			setTimeout( () => {
+				setVideoPlayerTemporaryHeightState( 'auto' );
+			}, 250 );
+			return;
 		}
 
 		if ( ! videoRatio ) {
@@ -186,12 +190,12 @@ export default function VideoPressPlayer( {
 				enable={ {
 					top: false,
 					bottom: false,
-					left: true,
+					left: false,
 					right: true,
 				} }
 				maxWidth="100%"
 				size={ { width: maxWidth } }
-				style={ { margin: 'auto' } }
+				style={ { marginRight: 'auto' } }
 				onResizeStop={ onBlockResize }
 				onResizeStart={ () => setVideoPlayerTemporaryHeightState( 'auto' ) }
 			>
@@ -212,10 +216,6 @@ export default function VideoPressPlayer( {
 								{ __( 'Loading…', 'jetpack-videopress-pkg' ) }
 							</div>
 						) }
-
-						<div className="jetpack-videopress-ghost-player">
-							<SandBox html={ html } />
-						</div>
 					</>
 				</div>
 			</ResizableBox>