+++ projects/packages/videopress/src/client/lib/video-tracks/types.ts
@@ -1,0 +1,11 @@
+import { TRACK_KIND_OPTIONS } from '.';
+
+type TrackKindOptionsProps = typeof TRACK_KIND_OPTIONS;
+export type trackKindOptionProps = TrackKindOptionsProps[ number ];
+
+export type UploadTrackDataProps = {
+	label: string;
+	srcLang: string;
+	kind: trackKindOptionProps;
+	tmpFile: File;
+};