+++ projects/plugins/jetpack/extensions/blocks/contact-form/child-blocks.js
@@ -1,6 +1,8 @@
+import { store as blockEditorStore } from '@wordpress/block-editor';
 import { createBlock, getBlockType } from '@wordpress/blocks';
 import { Circle, Path } from '@wordpress/components';
-import { Fragment } from '@wordpress/element';
+import { useSelect, useDispatch } from '@wordpress/data';
+import { Fragment, useEffect } from '@wordpress/element';
 import { __, _x } from '@wordpress/i18n';
 import { getIconColor } from '../../shared/block-icons';
 import renderMaterialIcon from '../../shared/render-material-icon';
@@ -11,8 +13,7 @@ import JetpackFieldMultiple from './components/jetpack-field-multiple';
 import JetpackFieldTextarea from './components/jetpack-field-textarea';
 
 const FieldDefaults = {
-	category: 'contact-form-fields',
-	parent: [ 'jetpack/contact-form' ],
+	category: 'contact-form',
 	supports: {
 		reusable: false,
 		html: false,
@@ -121,11 +122,45 @@ const FieldDefaults = {
 	example: {},
 };
 
+const validateFormWrapper = ( { attributes, clientId, name } ) => {
+	const FORM_BLOCK_NAME = 'jetpack/contact-form';
+	const BUTTON_BLOCK_NAME = 'jetpack/button';
+	const SUBMIT_BUTTON_ATTR = {
+		text: __( 'Submit', 'jetpack' ),
+		element: 'button',
+		lock: { remove: true },
+	};
+
+	// eslint-disable-next-line react-hooks/rules-of-hooks
+	const { replaceBlock } = useDispatch( blockEditorStore );
+
+	// eslint-disable-next-line react-hooks/rules-of-hooks
+	const parents = useSelect( select => {
+		return select( blockEditorStore ).getBlockParentsByBlockName( clientId, FORM_BLOCK_NAME );
+	} );
+
+	// eslint-disable-next-line react-hooks/rules-of-hooks
+	useEffect( () => {
+		if ( ! parents?.length ) {
+			replaceBlock(
+				clientId,
+				createBlock( FORM_BLOCK_NAME, {}, [
+					createBlock( name, attributes ),
+					createBlock( BUTTON_BLOCK_NAME, SUBMIT_BUTTON_ATTR ),
+				] )
+			);
+		}
+		// eslint-disable-next-line react-hooks/exhaustive-deps
+	}, [] );
+};
+
 const getFieldLabel = ( { attributes, name: blockName } ) => {
 	return null === attributes.label ? getBlockType( blockName ).title : attributes.label;
 };
 
 const editField = type => props => {
+	validateFormWrapper( props );
+
 	return (
 		<JetpackField
 			type={ type }
@@ -141,37 +176,48 @@ const editField = type => props => {
 	);
 };
 
-const editMultiField = type => props => (
-	<JetpackFieldMultiple
-		label={ getFieldLabel( props ) }
-		required={ props.attributes.required }
-		options={ props.attributes.options }
-		setAttributes={ props.setAttributes }
-		type={ type }
-		isSelected={ props.isSelected }
-		id={ props.attributes.id }
-		width={ props.attributes.width }
-	/>
-);
+const editMultiField = type => props => {
+	validateFormWrapper( props );
+
+	return (
+		<JetpackFieldMultiple
+			label={ getFieldLabel( props ) }
+			required={ props.attributes.required }
+			options={ props.attributes.options }
+			setAttributes={ props.setAttributes }
+			type={ type }
+			isSelected={ props.isSelected }
+			id={ props.attributes.id }
+			width={ props.attributes.width }
+		/>
+	);
+};
 
 export const childBlocks = [
 	{
 		name: 'field-text',
 		settings: {
 			...FieldDefaults,
-			title: __( 'Text', 'jetpack' ),
+			title: __( 'Text Input Field', 'jetpack' ),
 			description: __( 'When you need just a small amount of text, add a text input.', 'jetpack' ),
 			icon: renderMaterialIcon(
 				<Path fill={ getIconColor() } d="M4 9h16v2H4V9zm0 4h10v2H4v-2z" />
 			),
 			edit: editField( 'text' ),
+			attributes: {
+				...FieldDefaults.attributes,
+				label: {
+					type: 'string',
+					default: 'Text',
+				},
+			},
 		},
 	},
 	{
 		name: 'field-name',
 		settings: {
 			...FieldDefaults,
-			title: __( 'Name', 'jetpack' ),
+			title: __( 'Name Field', 'jetpack' ),
 			description: __(
 				'Introductions are important. Add an input for folks to add their name.',
 				'jetpack'
@@ -183,13 +229,20 @@ export const childBlocks = [
 				/>
 			),
 			edit: editField( 'text' ),
+			attributes: {
+				...FieldDefaults.attributes,
+				label: {
+					type: 'string',
+					default: 'Name',
+				},
+			},
 		},
 	},
 	{
 		name: 'field-email',
 		settings: {
 			...FieldDefaults,
-			title: __( 'Email', 'jetpack' ),
+			title: __( 'Email Field', 'jetpack' ),
 			keywords: [ __( 'e-mail', 'jetpack' ), __( 'mail', 'jetpack' ), 'email' ],
 			description: __( 'Want to reply to folks? Add an email address input.', 'jetpack' ),
 			icon: renderMaterialIcon(
@@ -199,14 +252,20 @@ export const childBlocks = [
 				/>
 			),
 			edit: editField( 'email' ),
+			attributes: {
+				...FieldDefaults.attributes,
+				label: {
+					type: 'string',
+					default: 'Email',
+				},
+			},
 		},
 	},
-
 	{
 		name: 'field-url',
 		settings: {
 			...FieldDefaults,
-			title: __( 'Website', 'jetpack' ),
+			title: __( 'URL Field', 'jetpack' ),
 			keywords: [ 'url', __( 'internet page', 'jetpack' ), 'link' ],
 			description: __( 'Add an address input for a website.', 'jetpack' ),
 			icon: renderMaterialIcon(
@@ -216,9 +275,15 @@ export const childBlocks = [
 				/>
 			),
 			edit: editField( 'url' ),
+			attributes: {
+				...FieldDefaults.attributes,
+				label: {
+					type: 'string',
+					default: 'URL',
+				},
+			},
 		},
 	},
-
 	{
 		name: 'field-date',
 		settings: {
@@ -236,13 +301,20 @@ export const childBlocks = [
 				/>
 			),
 			edit: editField( 'text' ),
+			attributes: {
+				...FieldDefaults.attributes,
+				label: {
+					type: 'string',
+					default: 'Date',
+				},
+			},
 		},
 	},
 	{
 		name: 'field-telephone',
 		settings: {
 			...FieldDefaults,
-			title: __( 'Phone Number', 'jetpack' ),
+			title: __( 'Phone Number Field', 'jetpack' ),
 			keywords: [
 				__( 'Phone', 'jetpack' ),
 				__( 'Cellular phone', 'jetpack' ),
@@ -256,13 +328,20 @@ export const childBlocks = [
 				/>
 			),
 			edit: editField( 'tel' ),
+			attributes: {
+				...FieldDefaults.attributes,
+				label: {
+					type: 'string',
+					default: 'Phone',
+				},
+			},
 		},
 	},
 	{
 		name: 'field-textarea',
 		settings: {
 			...FieldDefaults,
-			title: __( 'Message', 'jetpack' ),
+			title: __( 'Multi-line Text Field', 'jetpack' ),
 			keywords: [ __( 'Textarea', 'jetpack' ), 'textarea', __( 'Multiline text', 'jetpack' ) ],
 			description: __(
 				'Let folks speak their mind. This text box is great for longer responses.',
@@ -271,18 +350,29 @@ export const childBlocks = [
 			icon: renderMaterialIcon(
 				<Path fill={ getIconColor() } d="M21 11.01L3 11v2h18zM3 16h12v2H3zM21 6H3v2.01L21 8z" />
 			),
-			edit: props => (
-				<JetpackFieldTextarea
-					label={ getFieldLabel( props ) }
-					required={ props.attributes.required }
-					setAttributes={ props.setAttributes }
-					isSelected={ props.isSelected }
-					defaultValue={ props.attributes.defaultValue }
-					placeholder={ props.attributes.placeholder }
-					id={ props.attributes.id }
-					width={ props.attributes.width }
-				/>
-			),
+			edit: props => {
+				validateFormWrapper( props );
+
+				return (
+					<JetpackFieldTextarea
+						label={ props.attributes.label }
+						required={ props.attributes.required }
+						setAttributes={ props.setAttributes }
+						isSelected={ props.isSelected }
+						defaultValue={ props.attributes.defaultValue }
+						placeholder={ props.attributes.placeholder }
+						id={ props.attributes.id }
+						width={ props.attributes.width }
+					/>
+				);
+			},
+			attributes: {
+				...FieldDefaults.attributes,
+				label: {
+					type: 'string',
+					default: '',
+				},
+			},
 		},
 	},
 	{
@@ -298,17 +388,21 @@ export const childBlocks = [
 					d="M19 3H5c-1.1 0-2 .9-2 2v14c0 1.1.9 2 2 2h14c1.1 0 2-.9 2-2V5c0-1.1-.9-2-2-2zm0 16H5V5h14v14zM17.99 9l-1.41-1.42-6.59 6.59-2.58-2.57-1.42 1.41 4 3.99z"
 				/>
 			),
-			edit: props => (
-				<JetpackFieldCheckbox
-					label={ props.attributes.label } // label intentinally left blank
-					required={ props.attributes.required }
-					setAttributes={ props.setAttributes }
-					isSelected={ props.isSelected }
-					defaultValue={ props.attributes.defaultValue }
-					id={ props.attributes.id }
-					width={ props.attributes.width }
-				/>
-			),
+			edit: props => {
+				validateFormWrapper( props );
+
+				return (
+					<JetpackFieldCheckbox
+						label={ props.attributes.label } // label intentinally left blank
+						required={ props.attributes.required }
+						setAttributes={ props.setAttributes }
+						isSelected={ props.isSelected }
+						defaultValue={ props.attributes.defaultValue }
+						id={ props.attributes.id }
+						width={ props.attributes.width }
+					/>
+				);
+			},
 			attributes: {
 				...FieldDefaults.attributes,
 				label: {
@@ -322,7 +416,7 @@ export const childBlocks = [
 		name: 'field-consent',
 		settings: {
 			...FieldDefaults,
-			title: __( 'Consent', 'jetpack' ),
+			title: __( 'Terms Consent', 'jetpack' ),
 			keywords: [ __( 'Consent', 'jetpack' ) ],
 			description: __( 'Ask for consent', 'jetpack' ),
 			icon: renderMaterialIcon(
@@ -356,7 +450,9 @@ export const childBlocks = [
 					default: __( 'Can we send you an email from time to time?', 'jetpack' ),
 				},
 			},
-			edit: ( { attributes, isSelected, setAttributes } ) => {
+			edit: ( { attributes, clientId, isSelected, name, setAttributes } ) => {
+				validateFormWrapper( { attributes, clientId, name } );
+
 				const {
 					id,
 					width,
@@ -382,7 +478,7 @@ export const childBlocks = [
 		name: 'field-checkbox-multiple',
 		settings: {
 			...FieldDefaults,
-			title: __( 'Checkbox Group', 'jetpack' ),
+			title: __( 'Multiple Choice (Checkbox)', 'jetpack' ),
 			keywords: [ __( 'Choose Multiple', 'jetpack' ), __( 'Option', 'jetpack' ) ],
 			description: __( 'People love options. Add several checkbox items.', 'jetpack' ),
 			icon: renderMaterialIcon(
@@ -396,7 +492,7 @@ export const childBlocks = [
 				...FieldDefaults.attributes,
 				label: {
 					type: 'string',
-					default: 'Choose several',
+					default: 'Choose several options',
 				},
 			},
 		},
@@ -405,7 +501,7 @@ export const childBlocks = [
 		name: 'field-radio',
 		settings: {
 			...FieldDefaults,
-			title: __( 'Radio', 'jetpack' ),
+			title: __( 'Single Choice (Radio)', 'jetpack' ),
 			keywords: [ __( 'Choose', 'jetpack' ), __( 'Select', 'jetpack' ), __( 'Option', 'jetpack' ) ],
 			description: __(
 				'Inspired by radios, only one radio item can be selected at a time. Add several radio button items.',
@@ -425,7 +521,7 @@ export const childBlocks = [
 				...FieldDefaults.attributes,
 				label: {
 					type: 'string',
-					default: 'Choose one',
+					default: 'Choose one option',
 				},
 			},
 		},
@@ -434,7 +530,7 @@ export const childBlocks = [
 		name: 'field-select',
 		settings: {
 			...FieldDefaults,
-			title: __( 'Select', 'jetpack' ),
+			title: __( 'Dropdown Field', 'jetpack' ),
 			keywords: [
 				__( 'Choose', 'jetpack' ),
 				__( 'Dropdown', 'jetpack' ),
@@ -449,7 +545,7 @@ export const childBlocks = [
 				...FieldDefaults.attributes,
 				label: {
 					type: 'string',
-					default: 'Select one',
+					default: 'Select one option',
 				},
 			},
 		},