+++ projects/plugins/jetpack/extensions/shared/block-category.js
@@ -13,6 +13,17 @@ if ( ! isWpcom ) {
 	} );
 }
 
+// We're moving Form specific blocks to a new 'Forms' category
+// that should appear before the 'earn' and 'grow' categories
+setCategories( [
+	...getCategories().filter( ( { slug } ) => slug !== 'contact-form' ),
+	{
+		slug: 'contact-form',
+		title: __( 'Forms', 'jetpack' ),
+		icon: <JetpackLogo />,
+	},
+] );
+
 setCategories( [
 	...getCategories().filter( ( { slug } ) => slug !== 'earn' ),
 	// Add a Earn block category
@@ -31,16 +42,4 @@ setCategories( [
 		title: __( 'Grow', 'jetpack' ),
 		icon: <JetpackLogo />,
 	},
-] );
-
-setCategories( [
-	// Create a different category for the Contact Form fields
-	// to allow showing them first in the Block Library when a
-	// Form block is selected
-	{
-		slug: 'contact-form-fields',
-		title: __( 'Forms', 'jetpack' ),
-		icon: <JetpackLogo />,
-	},
-	...getCategories().filter( ( { slug } ) => slug !== 'contact-form-fields' ),
 ] );