+++ projects/plugins/jetpack/tests/php/modules/contact-form/test-class.grunion-contact-form.php
@@ -5,7 +5,7 @@
  * @package automattic/jetpack
  */
 
-require_jetpack_file( 'modules/contact-form/grunion-contact-form.php' );
+require_once JETPACK__PLUGIN_DIR . 'modules/contact-form/grunion-contact-form.php';
 
 /**
  * Test class for Grunion_Contact_Form
@@ -340,10 +340,10 @@ class WP_Test_Grunion_Contact_Form extends WP_UnitTestCase {
 		$submission = $feedback[0];
 		$email      = get_post_meta( $submission->ID, '_feedback_email', true );
 
-		$expected  = '<b>Name:</b> John Doe<br /><br />';
-		$expected .= '<b>Dropdown:</b> First option<br /><br />';
-		$expected .= '<b>Radio:</b> Second option<br /><br />';
-		$expected .= '<b>Text:</b> Texty text<br /><br />';
+		$expected  = '<div class="field-name">Name:</div> <div class="field-value">John Doe</div>';
+		$expected .= '<div class="field-name">Dropdown:</div> <div class="field-value">First option</div>';
+		$expected .= '<div class="field-name">Radio:</div> <div class="field-value">Second option</div>';
+		$expected .= '<div class="field-name">Text:</div> <div class="field-value">Texty text</div><br />';
 
 		$email_body = explode( PHP_EOL . PHP_EOL, $email['message'] );
 
@@ -393,10 +393,10 @@ class WP_Test_Grunion_Contact_Form extends WP_UnitTestCase {
 		$this->assertContains( 'john <john@example.com>', $args['to'] );
 		$this->assertEquals( 'Hello there!', $args['subject'] );
 
-		$expected  = '<b>Name:</b> John Doe<br /><br />';
-		$expected .= '<b>Dropdown:</b> First option<br /><br />';
-		$expected .= '<b>Radio:</b> Second option<br /><br />';
-		$expected .= '<b>Text:</b> Texty text<br /><br />';
+		$expected  = '<div class="field-name">Name:</div> <div class="field-value">John Doe</div>';
+		$expected .= '<div class="field-name">Dropdown:</div> <div class="field-value">First option</div>';
+		$expected .= '<div class="field-name">Radio:</div> <div class="field-value">Second option</div>';
+		$expected .= '<div class="field-name">Text:</div> <div class="field-value">Texty text</div><br />';
 
 		// Divides email by the first empty line.
 		$email_body = explode( PHP_EOL . PHP_EOL, $args['message'] );
@@ -736,7 +736,7 @@ class WP_Test_Grunion_Contact_Form extends WP_UnitTestCase {
 			'id'          => 'funID',
 		);
 
-		$expected_attributes = array_merge( $attributes, array( 'input_type' => 'url' ) );
+		$expected_attributes = array_merge( $attributes, array( 'input_type' => 'text' ) );
 		$this->assertValidField( $this->render_field( $attributes ), $expected_attributes );
 	}