+++ projects/packages/videopress/CHANGELOG.md
@@ -5,6 +5,92 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.1] - 2022-11-22
+### Added
+- Added VideoPress feedback link to the VideoPress block. [#27450]
+- VideoPress: add "Details panel" to v6 [#27428]
+- VideoPress: Add "learn how" chapters modal [#27438]
+- VideoPress: add block editor dependencies [#27489]
+- VideoPress: add dimensions panel to video block [#27520]
+- VideoPress: add rating control to Details panel [#27415]
+- VideoPress: check $products data before to pick the prices [#27504]
+- VideoPress: fix issue when setting video privacy [#27435]
+- VideoPress: implement "Allow download" control to v6 [#27420]
+- VideoPress: implement Privacy control [#27401]
+- VideoPress: improve when re-rendering the video player [#27546]
+- VideoPress: register and sync isPrivate block attribute [#27493]
+- VideoPress: Sync video tracks data in video block [#27495]
+- VideoPress: Updated the list of plans that have VideoPress included. [#27536]
+
+### Changed
+- Updated package dependencies. [#26069]
+- Updated package dependencies. [#26736]
+- Updated package dependencies. [#27043]
+- VideoPress: exposed the site purchases list on the client initial state, dropping the need of a request to the My Jetpack purchases endpoint. [#27533]
+- VideoPress: Remove extra resize circle on block [#27498]
+- VideoPress: show pricing based on sale coupons [#27535]
+- VideoPress: switch to v1.1/videos when requesting video data [#27488]
+- VideoPress: TypeScriptify code [#27419]
+
+### Fixed
+- VideoPress: Add backend verification on update routes when user is disconnected [#27455]
+- VideoPress: Add Jetpack Complete plan to list of valid purchased plans [#27500]
+- VideoPress: Disable actions when user is not connected or there is no connected site owner [#27402]
+- VideoPress: fix product plan [#27502]
+- VideoPress: Fix video getter function call [#27534]
+- VideoPress: Remove storage meter on free plan [#27549]
+- VideoPress: Set a static list of allowed video extensions allowed on VideoPress. [#27457]
+- VideoPress: Set the playback token on the video URL and the new poster URL when the video needs it. [#27404]
+
+## [0.8.0] - 2022-11-14
+### Added
+- VideoPress: add Color Panel component [#27381]
+- VideoPress: add file drop support after first video [#27297]
+- VideoPress: connect UI of settings section with the data handling. [#27350]
+- VideoPress: create endpoints to fetch and update the VideoPress settings. [#27293]
+- VideoPress: create selector/resolver machinery to fetch VideoPress site privacy setting from the backend. [#27321]
+- VideoPress: enable thumbnail selection from video frame on quick action [#27364]
+- VideoPress: exposed video_is_private flag on each video, taking into account the video privacy setting as well as the VideoPress site-wide privacy setting. [#27349]
+- VideoPress: mitigate video re-rendering flicker [#27305]
+
+### Changed
+- Updated package dependencies. [#27289]
+- VideoPress: allow keyboard navigation on video quick actions [#27378]
+- VideoPress: allow selection of multiple files [#27357]
+- VideoPress: introduce PlaybackControl component [#27303]
+- VideoPress: remove the fallback video [#27315]
+- VideoPress: rename Match video title with Dynamic color [#27385]
+- VideoPress: show Settings only when connected [#27361]
+- VideoPress: unify file drag and drop UI [#27348]
+
+### Fixed
+- VideoPress: fix an issue with private VideoPress videos timing out when script loading is delayed. [#27211]
+- VideoPress: change the resolver to rely on the new needsPlaybackToken flag when deciding if a video needs a playback token. [#27380]
+- VideoPress: disable automatic re-focus on video quick action popover close [#27338]
+- VideoPress: fix bug and return with select thumbnail from frame [#27309]
+
+## [0.7.0] - 2022-11-07
+### Added
+- Adding the new Video Chapter block. [#27240]
+- Videopress: add a notice and a modal informing the users how they can create chapters. [#27244]
+- VideoPress: add inspector controls to VideoPress chapters block. [#27242]
+- VideoPress: add Site Settings section. [#27290]
+- VideoPress: add video chapters style selector component. [#27237]
+- VideoPress: add VideoPress video block select control. [#27214]
+- VideoPress: introduce initial layout for chapters. [#27246]
+- VideoPress: scaffolding Video Chapters block. [#27241]
+- VideoPress: show message when there are no search results. [#27275]
+
+### Changed
+- Modified the VideoPress block's prompt to add a VideoChapters block so that the button functions as expected. [#27248]
+- VideoPress: improve the re-rendering process of the video player when editing video props. [#27192]
+- VideoPress: re-organize hooks to keep video data in sync. [#27227]
+
+### Fixed
+- VideoPress: change base number to 10 and total space value to 10^12 so we refer to it in terabytes instead of tebibytes, keeping consistency between marketing and product. [#27274]
+- VideoPress: fix privacy column icon. [#27277]
+- VideoPress: make sure media details are set before using them. [#27262]
+
 ## [0.6.5] - 2022-11-01
 ### Added
 - VideoPress: add getMediaToken() async helper [#27180]
@@ -383,6 +469,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Created empty package [#24952]
 
+[0.8.1]: https://github.com/Automattic/jetpack-videopress/compare/v0.8.0...v0.8.1
+[0.8.0]: https://github.com/Automattic/jetpack-videopress/compare/v0.7.0...v0.8.0
+[0.7.0]: https://github.com/Automattic/jetpack-videopress/compare/v0.6.5...v0.7.0
 [0.6.5]: https://github.com/Automattic/jetpack-videopress/compare/v0.6.4...v0.6.5
 [0.6.4]: https://github.com/Automattic/jetpack-videopress/compare/v0.6.3...v0.6.4
 [0.6.3]: https://github.com/Automattic/jetpack-videopress/compare/v0.6.2...v0.6.3